Create README.md
Write me the code for an SEO tool that has a beautiful form with beautiful buttons so that the user can enter the link to the desired site. After entering the code, it will analyze all the SEO information of the site and present it in a table or beautiful form.